+++ Flashcards.wkktoria/Flashcards.wkktoria/Managers/AllStackManager.cs
@@ -3,7 +3,7 @@ using Flashcards.wkktoria.Services;
 using Flashcards.wkktoria.UserInteractions;
 using Flashcards.wkktoria.UserInteractions.Helpers;
 
-namespace Flashcards.wkktoria.Managers.Helpers;
+namespace Flashcards.wkktoria.Managers;
 
 internal class AllStackManager
 {
