+++ Flashcards.wkktoria/Flashcards.wkktoria/Managers/SessionManager.cs
@@ -1,4 +1,5 @@
 using Flashcards.wkktoria.Controllers;
+using Flashcards.wkktoria.Managers.Helpers;
 using Flashcards.wkktoria.Models;
 using Flashcards.wkktoria.Services;
 using Flashcards.wkktoria.UserInteractions;
@@ -20,30 +21,12 @@ internal class SessionManager
 
     internal void Run()
     {
-        Choose();
         Console.Clear();
 
-        _sessionController.ShowAll(_currentStack!.Id);
-    }
-
-    private void Choose()
-    {
-        Console.Clear();
-
-        var stacks = _stackService.GetAll();
-
-        if (stacks.Any())
+        if (_stackService.GetAll().Any())
         {
-            TableVisualisation.ShowStacksTable(stacks);
-
-            do
-            {
-                var id = UserInput.GetNumberInput("Enter id of stack to see study sessions.");
-                var stack = _stackService.GetByDtoId(id);
-                _currentStack = stack;
-
-                if (_currentStack == null) UserOutput.ErrorMessage("No stack found.");
-            } while (_currentStack == null);
+            _currentStack = StackHelper.Choose(_stackService);
+            _sessionController.ShowAll(_currentStack!.Id);
         }
         else
         {
